Move LeftSideButton fetch to API Add close option for LeftSideButton
