+++ app/components/homePage/LeftSideButton.tsx
@@ -2,10 +2,11 @@
 import { LeftSideButtonData } from "@/types/types";
 //---Framework---//
 import { useEffect, useState, FC } from "react";
-import client from "../../../sanity/lib/client";
+import axios from "axios";
 import Link from "next/link";
 //---Fonts---//
 import { playfairDisplay } from "@/app/fonts/fonts";
+import { X } from "@phosphor-icons/react";
 
 const LeftSideButton: FC = () => {
 	const [leftSideButtonData, setLeftSideButtonData] =
@@ -14,10 +15,13 @@ const LeftSideButton: FC = () => {
 
 	useEffect(() => {
 		const fetchData = async () => {
-			const query = `*[_type == "sideButton" && enabled == true][0]`;
-			const data = await client.fetch(query);
-			if (data) {
-				setLeftSideButtonData(data);
+			try {
+				const response = await axios.get("/api/layout/left-side-button");
+				if (response.data) {
+					setLeftSideButtonData(response.data);
+				}
+			} catch (error) {
+				console.error("Error fetching side button data:", error);
 			}
 		};
 
@@ -28,22 +32,35 @@ const LeftSideButton: FC = () => {
 		return null;
 	}
 	return (
-		<Link href={leftSideButtonData.link} passHref>
+		<div className="relative">
+			<Link href={leftSideButtonData.link} passHref>
+				<button
+					className="fixed left-[.1rem] top-[38%] lg:top-1/2 -translate-y-1/2 z-[60] flex flex-col justify-center items-center text-white p-5 rounded-md text-center font-bold text-xs sm:text-sm md:text-md lg:text-lg border-[.1rem] border-black shadow-lg hover:bg-secondary"
+					id="sideButton"
+					style={{
+						writingMode: "vertical-rl",
+						transform: "rotate(180deg)",
+						backgroundColor: leftSideButtonData.backgroundColor?.hex || "white",
+						color: leftSideButtonData.textColor?.hex || "black",
+					}}
+				>
+					<span className={`tracking-widest ${playfairDisplay.className}`}>
+						{leftSideButtonData.text}
+					</span>
+				</button>
+			</Link>
 			<button
-				className="fixed left-[.1rem] top-[38%] lg:top-1/2 -translate-y-1/2 z-[60] flex flex-col justify-center items-center text-white p-5 rounded-md text-center font-bold text-xs sm:text-sm md:text-md lg:text-lg border-[.1rem] border-black shadow-lg hover:bg-secondary"
-				id="sideButton"
+				className="fixed top-[38%] lg:top-[50.5%] left-[.2rem] -translate-y-1/2 text-sm text-black p-1 z-[400] hover:text-red-500 rounded-full hover:bg-red-50"
+				onClick={() => setIsVisible(false)}
 				style={{
 					writingMode: "vertical-rl",
 					transform: "rotate(180deg)",
-					backgroundColor: leftSideButtonData.backgroundColor?.hex || "white",
 					color: leftSideButtonData.textColor?.hex || "black",
 				}}
 			>
-				<span className={`tracking-widest ${playfairDisplay.className}`}>
-					{leftSideButtonData.text}
-				</span>
+				<X className="w-4 h-4" />
 			</button>
-		</Link>
+		</div>
 	);
 };
